+++ tests/test_label_maximum_extension_ratio_map.py
@@ -1,22 +1,22 @@
 import pyclesperanto_prototype as cle
 import numpy as np
 
-def test_label_pixel_count_map():
+def test_label_label_maximum_extension_ratio_map_2d():
 
     labels = cle.push_zyx(np.asarray([
         [1, 1, 2],
-        [1, 0, 0],
+        [1, 0, 2],
         [3, 3, 0]
     ]))
 
     reference = cle.push_zyx(np.asarray([
-        [3, 3, 1],
-        [3, 0, 0],
-        [2, 2, 0]
+        [1.1396203, 1.1396203, 1],
+        [1.1396203, 0, 1],
+        [1, 1, 0]
     ]
     ))
 
-    result = cle.label_pixel_count_map(labels)
+    result = cle.label_maximum_extension_ratio_map(labels)
 
     a = cle.pull_zyx(result)
     b = cle.pull_zyx(reference)
@@ -24,16 +24,16 @@ def test_label_pixel_count_map():
     print(a)
     print(b)
 
-    assert (np.array_equal(a, b))
+    assert (np.allclose(a, b, 0.001))
 
 
-def test_label_pixel_count_map_3d():
+def test_label_label_maximum_extension_ratio_map_3d():
 
     labels = cle.push_zyx(np.asarray([
         [
             [1, 1, 2],
         ], [
-            [1, 0, 0],
+            [1, 0, 2],
         ], [
             [3, 3, 0]
         ]
@@ -41,16 +41,16 @@ def test_label_pixel_count_map_3d():
 
     reference = cle.push_zyx(np.asarray([
         [
-            [3, 3, 1],
+            [1.1396203, 1.1396203, 1],
         ], [
-            [3, 0, 0],
+            [1.1396203, 0, 1],
         ], [
-            [2, 2, 0]
+            [1, 1, 0]
         ]
     ]
     ))
 
-    result = cle.label_pixel_count_map(labels)
+    result = cle.label_maximum_extension_ratio_map(labels)
 
     a = cle.pull_zyx(result)
     b = cle.pull_zyx(reference)
@@ -58,7 +58,7 @@ def test_label_pixel_count_map_3d():
     print(a)
     print(b)
 
-    assert (np.array_equal(a, b))
+    assert (np.allclose(a, b, 0.001))