+++ tests/test_statistics_of_background_and_labelled_pixels.py
@@ -46,17 +46,16 @@ def test_statistics_of_background_and_labelled_pixels():
         [0.6666667,    2.3333333,    3.],
         #     SUM_INTENSITY(13),
         [2.,          7.,          9.],
-
         #     STANDARD_DEVIATION_INTENSITY(14),
         [0.47140452, 0.4714045, 0.8164966],
         #     PIXEL_COUNT(15),
         [3.,    3.,    3.],
         #     SUM_INTENSITY_TIMES_X(16),
-        [0.,          0.,          0.],# Todo; not supported yet
+        [1., 12., 11.],
         #     SUM_INTENSITY_TIMES_Y(17),
-        [0.,    0.,    0.],# Todo; not supported yet
-        #     SUM_INTENSITY_TIMES_Z(18),
-        [0.,          0.,          0.],# Todo; not supported yet
+        [1., 5., 18.],
+        #     SUM_INTENSITY_TIMES_Y(17),
+        [0., 0., 0.],
         #     MASS_CENTER_X(19),
         [0.5,    1.7142857,    1.2222222],
         #     MASS_CENTER_Y(20),
@@ -64,11 +63,11 @@ def test_statistics_of_background_and_labelled_pixels():
         #     MASS_CENTER_Z(21),
         [0.,    0.,    0.],
         #     SUM_X(22),
-        [0.,          0.,          0.],# Todo; not supported yet
+        [1., 5., 3.],
         #     SUM_Y(23),
-        [0.,     0.,    0.],# Todo; not supported yet
+        [1., 2., 6.],
         #     SUM_Z(24),
-        [0.,          0.,          0.],# Todo; not supported yet
+        [0., 0., 0.],
         #     CENTROID_X(25),
         [0.33333334,    1.6666666,    1.],
         #     CENTROID_Y(26),
@@ -76,13 +75,13 @@ def test_statistics_of_background_and_labelled_pixels():
         #     CENTROID_Z(27),
         [0.,    0.,    0.],
         #     SUM_DISTANCE_TO_MASS_CENTER(28),
-        [0.,          0.,          0.],# Todo; not supported yet
+        [2.1213202, 1.9426796, 2.2222223],
         #     MEAN_DISTANCE_TO_MASS_CENTER(29),
-        [0.,    0.,    0.],# Todo; not supported yet
+        [0.70710677, 0.6475599, 0.7407408],
         #     MAX_DISTANCE_TO_MASS_CENTER(30),
-        [0.,          0.,          0.],# Todo; not supported yet
+        [0.70710677, 0.7693094, 1.2222222],
         #     MAX_MEAN_DISTANCE_TO_MASS_CENTER_RATIO(31),
-        [0.,    0.,    0.],# Todo; not supported yet
+        [1., 1.1880127, 1.6499999],
         #     SUM_DISTANCE_TO_CENTROID(32),
         [1.9621165, 1.9621165, 2.],
         #     MEAN_DISTANCE_TO_CENTROID(33),
@@ -103,7 +102,7 @@ def test_statistics_of_background_and_labelled_pixels():
     print(a)
     print(b)
 
-    assert (np.array_equal(a, b))
+    assert (np.allclose(a, b, 0.0001))
 
 
 def test_statistics_of_background_and_labelled_pixels_3d():
@@ -128,74 +127,74 @@ def test_statistics_of_background_and_labelled_pixels_3d():
     ]))
 
     reference = cle.push_zyx(np.asarray([
-        #IDENTIFIER(0),
-        [0.,         1.,          2.],
+        # IDENTIFIER(0),
+        [0., 1., 2.],
         # BOUNDING_BOX_X(1),
-        [0.,          1.,          0.],
+        [0., 1., 0.],
         #     BOUNDING_BOX_Y(2),
-        [0.,          0.,          0.],
+        [0., 0., 0.],
         #     BOUNDING_BOX_Z(3),
         [0., 0., 2.],
 
         #     BOUNDING_BOX_END_X(4),
-        [1.,    2.,    2.],
+        [1., 2., 2.],
         #     BOUNDING_BOX_END_Y(5),
-        [0.,    0.,    0.],
+        [0., 0., 0.],
         #     BOUNDING_BOX_END_Z(6),
         [1., 1., 2.],
 
         #     BOUNDING_BOX_WIDTH(7),
-        [2.,         2.,          3.],
+        [2., 2., 3.],
         #     BOUNDING_BOX_HEIGHT(8),
-        [1.,         1.,          1.],
+        [1., 1., 1.],
         #     BOUNDING_BOX_DEPTH(9),
         [2., 2., 1.],
 
         #     MINIMUM_INTENSITY(10),
-        [0.,    2.,    2.],
+        [0., 2., 2.],
         #     MAXIMUM_INTENSITY(11),
-        [1.,          3.,          4.],
+        [1., 3., 4.],
         #     MEAN_INTENSITY(12)
-        [0.6666667,    2.3333333,    3.],
+        [0.6666667, 2.3333333, 3.],
         #     SUM_INTENSITY(13),
-        [2.,          7.,          9.],
+        [2., 7., 9.],
 
         #     STANDARD_DEVIATION_INTENSITY(14),
         [0.47140452, 0.4714045, 0.8164966],
         #     PIXEL_COUNT(15),
-        [3.,    3.,    3.],
+        [3., 3., 3.],
         #     SUM_INTENSITY_TIMES_X(16),
-        [0.,          0.,          0.],# Todo; not supported yet
+        [1., 12., 11.],
         #     SUM_INTENSITY_TIMES_Y(17),
-        [0.,    0.,    0.],# Todo; not supported yet
+        [0., 0., 0.],
         #     SUM_INTENSITY_TIMES_Z(18),
-        [0.,          0.,          0.],# Todo; not supported yet
+        [1., 5., 18.],
         #     MASS_CENTER_X(19),
-        [0.5,    1.7142857,    1.2222222],
+        [0.5, 1.7142857, 1.2222222],
         #     MASS_CENTER_Y(20),
-        [0.,    0.,    0.],
+        [0., 0., 0.],
         #     MASS_CENTER_Z(21),
-        [0.5,         0.71428573,  2.],
+        [0.5, 0.71428573, 2.],
         #     SUM_X(22),
-        [0.,          0.,          0.],# Todo; not supported yet
+        [1., 5., 3.],
         #     SUM_Y(23),
-        [0.,     0.,    0.],# Todo; not supported yet
+        [0., 0., 0.],
         #     SUM_Z(24),
-        [0.,          0.,          0.],# Todo; not supported yet
+        [1., 2., 6.],
         #     CENTROID_X(25),
-        [0.33333334,    1.6666666,    1.],
+        [0.33333334, 1.6666666, 1.],
         #     CENTROID_Y(26),
-        [0.,    0.,    0.],
+        [0., 0., 0.],
         #     CENTROID_Z(27),
-        [0.33333334,  0.6666667,   2.],
+        [0.33333334, 0.6666667, 2.],
         #     SUM_DISTANCE_TO_MASS_CENTER(28),
-        [0.,          0.,          0.],# Todo; not supported yet
+        [2.1213202, 1.9426795, 2.2222223],
         #     MEAN_DISTANCE_TO_MASS_CENTER(29),
-        [0.,    0.,    0.],# Todo; not supported yet
+        [0.70710677, 0.6475599, 0.7407408],
         #     MAX_DISTANCE_TO_MASS_CENTER(30),
-        [0.,          0.,          0.],# Todo; not supported yet
+        [0.70710677, 0.7693094, 1.2222222],
         #     MAX_MEAN_DISTANCE_TO_MASS_CENTER_RATIO(31),
-        [0.,    0.,    0.],# Todo; not supported yet
+        [1., 1.1880127, 1.6499999],
         #     SUM_DISTANCE_TO_CENTROID(32),
         [1.9621165, 1.9621165, 2.],
         #     MEAN_DISTANCE_TO_CENTROID(33),
@@ -216,10 +215,10 @@ def test_statistics_of_background_and_labelled_pixels_3d():
     print(a)
     print(b)
 
-    assert (np.array_equal(a, b))
+    assert (np.allclose(a, b, 0.0001))
 
 
-def test_statistics_of_background_and_labelled_pixels_compare_to_clij2():
+def test_statistics_of_background_and_labelled_pixels_gpu_compare_to_clij2():
     from skimage.io import imread
 
     blobs = cle.push_zyx(imread('data/mini_blobs.tif'))
@@ -262,15 +261,10 @@ def test_statistics_of_background_and_labelled_pixels_compare_to_clij2():
         cle.STATISTICS_ENTRY.CENTROID_X.value,
         cle.STATISTICS_ENTRY.CENTROID_Y.value,
         cle.STATISTICS_ENTRY.CENTROID_Z.value,
-        # Errors:
         cle.STATISTICS_ENTRY.SUM_DISTANCE_TO_CENTROID.value,
         cle.STATISTICS_ENTRY.MEAN_DISTANCE_TO_CENTROID.value,
         cle.STATISTICS_ENTRY.MAX_DISTANCE_TO_CENTROID.value,
-        cle.STATISTICS_ENTRY.MAX_MEAN_DISTANCE_TO_CENTROID_RATIO.value
-    ]
-
-    # TODO: not yet implemented:
-    not_yet_implemeneted_columns = [
+        cle.STATISTICS_ENTRY.MAX_MEAN_DISTANCE_TO_CENTROID_RATIO.value,
         cle.STATISTICS_ENTRY.SUM_INTENSITY_TIMES_X.value,
         cle.STATISTICS_ENTRY.SUM_INTENSITY_TIMES_Y.value,
         cle.STATISTICS_ENTRY.SUM_INTENSITY_TIMES_Z.value,
@@ -281,8 +275,8 @@ def test_statistics_of_background_and_labelled_pixels_compare_to_clij2():
         cle.STATISTICS_ENTRY.MEAN_DISTANCE_TO_MASS_CENTER.value,
         cle.STATISTICS_ENTRY.MAX_DISTANCE_TO_MASS_CENTER.value,
         cle.STATISTICS_ENTRY.MAX_MEAN_DISTANCE_TO_MASS_CENTER_RATIO.value,
-
     ]
+
     for column_index in columns_to_test:
         reference_values = clij_reference_table[:, column_index]
         values = table[:, column_index]